Add Jest and refactor server
