+++ package.json
@@ -4,7 +4,7 @@
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
     "create-admin": "node criar-admin.js",
     "clean-users": "node limpar-usuarios.js"
   },
@@ -28,5 +28,8 @@
     "winston": "^3.17.0",
     "ws": "^8.18.2",
     "xlsx": "^0.18.5"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0"
   }
 }
